fixing user sign up
