+++ routes/api/deleteGroups.js
