+++ routes/api/viewGroups.js
